+++ handler/ingest.go
@@ -1,4 +1,4 @@
-package syslog
+package handler
 
 import (
 	"fmt"
